0-read_textfile.c

+++ 0x15-file_io/0-read_textfile.c
@@ -50,5 +50,5 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	free(buff);
 	close(fpo);
 
-	return (rnum);
+	return ((ssize_t)rnum);
 }
